+++ resources/views/menuitem/index.blade.php
@@ -5,12 +5,14 @@
         @csrf
         <x-laraknife.panels.index title="{{ __('Menu Items') }}">
             <x-laraknife.panels.filter legend="{{ $pagination->legendText() }}">
-                <x-laraknife.forms.string position="alone" name="text" label="Name" value="{{ $context->valueOf('text') }}"
+                <x-laraknife.forms.string position="first" name="text" label="Name" value="{{ $context->valueOf('text') }}"
                     width2="4" />
+                <x-laraknife.forms.combobox position="last" name="section" label="Section" :options="$optionsSection"
+                    class="lkn-autoupdate" width2="4" />
             </x-laraknife.panels.filter>
             <x-laraknife.forms.set-position position="first" />
-            <x-laraknife.buttons.button name="btnNew" label="{{ __('New')}}" width1="0" width2="6"  />
-            <x-laraknife.buttons.button name="btnAssign" label="{{ __('Assign Roles')}}" width1="0" width2="6"  />
+            <x-laraknife.buttons.button name="btnNew" label="{{ __('New') }}" width1="0" width2="6" />
+            <x-laraknife.buttons.button name="btnAssign" label="{{ __('Assign Roles') }}" width1="0" width2="6" />
             <x-laraknife.forms.set-position position="last" />
             <x-laraknife.panels.sortable-table :context="$context" :pagination="$pagination">
                 <thead>